Bluetooth: Tester: Fix Scheduler Action set
Fix inncorrectly declared month parameter in the Scheduler Action Set
struct in the bttester.h.

+++ tests/bluetooth/tester/src/bttester.h
@@ -621,7 +621,7 @@ struct mesh_scheduler_action_set {
 	uint8_t ack;
 	uint8_t index;
 	uint8_t year;
-	uint8_t month;
+	uint16_t month;
 	uint8_t day;
 	uint8_t hour;
 	uint8_t minute;
